+++ R/OptimizerBatchCmaes.R
@@ -4,28 +4,30 @@
 #' @name mlr_optimizers_cmaes
 #'
 #' @description
-#' `OptimizerBatchCmaes` class that implements CMA-ES. Calls [adagio::pureCMAES()]
-#' from package \CRANpkg{adagio}. The algorithm is typically applied to search
-#' space dimensions between three and fifty. Lower search space dimensions might
-#' crash.
+#' `OptimizerBatchCmaes` class that implements CMA-ES.
+#' Calls `cma_es()` from package \CRANpkg{cmaes}.
+#' The algorithm is typically applied to search space dimensions between three and fifty.
+#' Lower search space dimensions might crash.
 #'
 #' @templateVar id cmaes
 #' @template section_dictionary_optimizers
 #'
 #' @section Parameters:
 #' \describe{
-#' \item{`sigma`}{`numeric(1)`}
 #' \item{`start_values`}{`character(1)`\cr
-#' Create `"random"` start values or based on `"center"` of search space?
-#' In the latter case, it is the center of the parameters before a trafo is applied.
-#' If set to `"custom"`, the start values can be passed via the `start` parameter.}
+#'   Create `"random"` start values or based on `"center"` of search space?
+#'   In the latter case, it is the center of the parameters before a trafo is applied.
+#'   If set to `"custom"`, the start values can be passed via the `start` parameter.}
 #' \item{`start`}{`numeric()`\cr
-#' Custom start values. Only applicable if `start_values` parameter is set to `"custom"`.}
+#'   Custom start values.
+#'   Only applicable if `start_values` parameter is set to `"custom"`.}
 #' }
 #'
-#' For the meaning of the control parameters, see [adagio::pureCMAES()]. Note
-#' that we have removed all control parameters which refer to the termination of
-#' the algorithm and where our terminators allow to obtain the same behavior.
+#' For the meaning of the control parameters, see `cma_es()`.
+#' The parameters `maxit`, `stopfitness` and `stop.tolx` can be used additionally to our terminators.
+#' The default values of `maxit` is `100 * D^2` where `D` is the number of dimensions of the search space.
+#' The `stop.tolx` parameter stops when the step size is smaller than `1e-12 * sigma`.
+#' The `vectorized` parameter is always set to `TRUE`.
 #'
 #' @template section_progress_bars
 #'
@@ -72,9 +74,26 @@ OptimizerBatchCmaes = R6Class("OptimizerBatchCmaes",
     #' Creates a new instance of this [R6][R6::R6Class] class.
     initialize = function() {
       param_set = ps(
-        sigma = p_dbl(default = 0.5),
-        start_values = p_fct(default = "random", levels = c("random", "center", "custom")),
-        start = p_uty(default = NULL, depends = start_values == "custom")
+        fnscale       = p_dbl(default = 1),
+        maxit         = p_int(lower  = 1L),
+        stopfitness   = p_dbl(default = -Inf),
+        keep.best     = p_lgl(default = TRUE),
+        sigma         = p_uty(default = 0.5),
+        mu            = p_int(lower = 1L),
+        lambda        = p_int(lower = 1L),
+        weights       = p_uty(),
+        damps         = p_dbl(),
+        cs            = p_dbl(),
+        ccum          = p_dbl(),
+        ccov.1        = p_dbl(lower = 0),
+        ccov.mu       = p_dbl(lower = 0),
+        diag.sigma    = p_lgl(default = FALSE),
+        diag.eigen    = p_lgl(default = FALSE),
+        diag.pop      = p_lgl(default = FALSE),
+        diag.value    = p_lgl(default = FALSE),
+        stop.tolx     = p_dbl(), # undocumented stop criterion
+        start_values  = p_fct(default = "random", levels = c("random", "center", "custom")),
+        start         = p_uty(default = NULL, depends = start_values == "custom")
       )
       param_set$values$start_values = "random"
       param_set$values$start = NULL
@@ -84,7 +103,7 @@ OptimizerBatchCmaes = R6Class("OptimizerBatchCmaes",
         param_set = param_set,
         param_classes = "ParamDbl",
         properties = "single-crit",
-        packages = "adagio",
+        packages = "cmaes",
         label = "Covariance Matrix Adaptation Evolution Strategy",
         man = "bbotk::mlr_optimizers_cmaes"
       )
@@ -94,29 +113,33 @@ OptimizerBatchCmaes = R6Class("OptimizerBatchCmaes",
   private = list(
     .optimize = function(inst) {
       pv = self$param_set$values
+      start_values = pv$start_values
+      start = pv$start
 
-      if (pv$start_values == "custom") {
-        pv$par = pv$start
-        pv$start_values = NULL
-        pv$start = NULL
-      } else {
-        pv$par = search_start(inst$search_space, type = pv$start_values)
-        pv$start_values = NULL
-        pv$start = NULL
-      }
 
-      pv$stopeval = .Machine$integer.max # make sure pureCMAES does not stop
-      pv$stopfitness = -Inf
+      par = if (pv$start_values == "custom") set_names(start, inst$search_space$ids()) else search_start(inst$search_space, type = start_values)
 
-      if (length(pv$par) < 2L) {
+      if (length(par) < 2L) {
         warning("CMA-ES is typically applied to search space dimensions between three and fifty. A lower search space dimension might crash.")
       }
 
-      invoke(adagio::pureCMAES,
-        fun = inst$objective_function,
+      control = pv[names(pv) %nin% c("start_values", "start")]
+      control$vectorized = TRUE
+
+      wrapper = function(xmat, inst) {
+        xdt = as.data.table(t(xmat))
+        res = inst$eval_batch(xdt)
+        y = res[, inst$objective$codomain$target_ids, with = FALSE][[1]]
+        y * inst$objective_multiplicator
+      }
+
+      invoke(cmaes::cma_es,
+        par = par,
+        fn = wrapper,
         lower = inst$search_space$lower,
         upper = inst$search_space$upper,
-        .args = pv)
+        control = control,
+        inst = inst)
     }
   )
 )